Implemented exercises page. Now Version 2.0
Used the Timeline.js plugin to create timelines for each of the separate
modules on the timetable page. NB, if the modules have no exercises then
they're not picked off.

Parser was a hasty affair, so not 100% on whether the dates are all
absolutely perfect though they certainly seem to be on the ball.

TODO - Come up with some way of reducing the UI on the exercises page,
       ie, have selective viewing of modules. Looking for ideas here...
     - Failing first point, just improve performance when first entering
       the exercises page.
     - Deal with bug on the timeline.js end.
     - Prettify it somehow. Don't know why, but it looks somehow
       --bare-- at the moment.
     - Implement a colored shading in the timeline.js where module that
       are currently active will show with a red shadow, those completed
       with a green and those up coming with a blue. Think this should
       go some way toward highlighting the important info.