Removed DS_Store's
Got rid of all the nasty DS_Store files, added a line to .gitignore
and created a symlink to the local timeline coffee file in the views
folder. Symlink for now so any local changes I make can be seen, using
the scraper to provide code so I can fine tune the timeline code.

Web.rb also updated to supply the file.