Fixed date detection bug and implemented one tl.
Date detection was incorrect when trying to pick up the first and last
dates of a term. Fixed this by using parseFloat instead of parseInt for
the date text (not sure why this was a problem) and fixed the detection
of the months to avoid cate's nasty habit of truncating month names
horrible.

Bit of a patch up seeing as it required a single case for the 'J' month,
ie June, but should work pretty well.