Implemented given view and sorted tl links
Implemented the scraping of the given pages and a bootstrap modal view
for the links. Peter's probably gonna want to clean this the hell up but
it should be pretty solid.

Timeline had big problem as to how to access the functionality of the
modal views etc that had already been implemented. Fixed that by
changing what's sent in with the moment- if it receives a string then
it'll go ahead and set the href property, and prevent propagation on the
click event. If however the value in the moment for the given key is an
object, then assume it's a jquery element and send it a click trigger,
activating all the functionality within the sealed Classy script without
any code duplication and/or unnecessary function exposure.

Quite happy with this, just need to figure out how to integrate this
into other pages (ie the dashboard) without having to copy in some
pointless stuff.

TODO - Implement on the dashboard page, finally.
     - Work on scraping the handin page and create a nice modal view for
       the content entry. Could make it look so much nicer than it does,
       and on doing that we will have covered all the Givens, Notes,
       Specs and Handins. Last thing to be done for the exercise page
       scraping wise.
     - Sort out ideas for how to integrate the timeline into the
       existing timetable page so as to provide a viewing option instead
       of something forced.
     - Work on some colour coding! Get the tables to colour code the
       exercises that are in over the next week in orange, and those
       that are in tomorrow/today in red. Gonna need to add
       functionality to pick off the due TIME also, seeing as that's
       pretty important.
     - Better dashboard, but that's for a different branch.

WISHLIST - Get interation between the exercise tables and the timelines,
           ie click on exercise, timeline counterpart expands. Would be
           pretty cool.